[chore] classroom画面のコンポーネント化
